refactor project actions
