--- src/.local
@@ -1,3 +1,0 @@
-acc 1 :
-tribeti@yopmail.com
-xIfZjplt710Q
