+++ src/utils/board-access.ts
@@ -185,17 +185,5 @@ export async function syncPrimaryAssignee(
   }
 }
 
-export function validateString(
-  value: string,
-  fieldName: string,
-  maxLength: number = 500,
-): string {
-  const trimmed = value.trim();
-  if (trimmed.length === 0) {
-    throw new Error(`${fieldName} is required.`);
-  }
-  if (trimmed.length > maxLength) {
-    throw new Error(`${fieldName} must be ${maxLength} characters or less.`);
-  }
-  return trimmed;
-}
+// Re-export from dedicated utility for backward compatibility
+export { validateString } from "@/utils/validate-string";
